+++ ft_memcmp.c
@@ -6,7 +6,7 @@
 /*   By: sadawi <sadawi@student.hive.fi>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/10/15 14:57:01 by sadawi            #+#    #+#             */
-/*   Updated: 2019/10/28 19:02:38 by sadawi           ###   ########.fr       */
+/*   Updated: 2019/11/06 13:46:07 by sadawi           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,10 +14,10 @@
 
 int	ft_memcmp(const void *s1, const void *s2, size_t n)
 {
-	int i;
+	size_t i;
 
 	i = 0;
-	while (i < (int)n)
+	while (i < n)
 	{
 		if (((unsigned char*)s1)[i] != ((unsigned char*)s2)[i])
 			return (((unsigned char*)s1)[i] - ((unsigned char*)s2)[i]);
