+++ ft_strncat.c
@@ -6,7 +6,7 @@
 /*   By: sadawi <sadawi@student.hive.fi>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/10/17 15:33:02 by sadawi            #+#    #+#             */
-/*   Updated: 2019/10/29 13:37:33 by sadawi           ###   ########.fr       */
+/*   Updated: 2019/11/06 13:52:15 by sadawi           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,14 +14,14 @@
 
 char	*ft_strncat(char *s1, const char *s2, size_t n)
 {
-	int i;
-	int j;
+	size_t i;
+	size_t j;
 
 	i = 0;
 	j = 0;
 	while (s1[i])
 		i++;
-	while (s2[j] && (int)n - j > 0)
+	while (s2[j] && n - j > 0)
 	{
 		s1[i + j] = s2[j];
 		j++;
